Moodify_AI/requirements.txt
aws-cdk-lib==2.134.0
constructs>=10.0.0,<11.0.0
requests~=2.31.0
python-dotenv
black
pycryptodome~=3.21.0
ldap3==2.9.1
pyasn1>=0.4.6